Move around enum declarations to reduce circular dependencies
